Move TagManager from widgets module to views
