+++ qwatson/views/activitydialog.py
@@ -25,6 +25,49 @@ from qwatsion.widgets.layout import ColoredFrame
 from qwatson.utils import icons
 
 
+class TagManager(QWidget):
+    def __init__(self, tags=[], parent=None):
+        super(TagManager, self).__init__(parent)
+        self.setup()
+        self.set_tags(tags)
+
+    def setup(self):
+        """Setup the widget with the provided arguments."""
+        self.tag_edit = TagLineEdit()
+        self.tag_edit.setPlaceholderText("Tags (comma separated)")
+
+        # ---- Setup the layout
+
+        layout = QGridLayout(self)
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.addWidget(self.tag_edit)
+
+    @property
+    def tags(self):
+        """Return the list of tags entered in the tag edit line."""
+        return self.tag_edit.tags
+
+    def set_tags(self, tags):
+        """Add a the tag list to the tag edit line."""
+        self.tag_edit.set_tags(tags)
+
+
+class TagLineEdit(QLineEdit):
+    """A lineedit to show and edit tags."""
+    def __init__(self, parent=None):
+        super(TagLineEdit, self).__init__(parent)
+
+    @property
+    def tags(self):
+        """Return the list of tags."""
+        tags = self.text().split(',')
+        return sorted(set(tag.strip() for tag in tags))
+
+    def set_tags(self, tags):
+        """Add a the tag list."""
+        self.setText(', '.join(tags))
+
+
 class ProjectManager(QWidget):
     """
     A toolbar composed of a combobox to show the existing projects and
